spearbit-71: add stricter validations to transfer/withdrawal/deposit
