+++ scripts/decode_calldata.py
@@ -1,0 +1,22 @@
+import json
+from web3 import Web3
+
+ABI_PATH = 'GRVTExchangeTest.json'
+CALLDATA = '0xe7cacbe80000000000000000000000000000000000000000000000000000034630b8a000000000000000000000000000000000000000000000000000000000000000001700000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000006c000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000001f4000000000000000000000000000000000000000000000000000000000000012c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001e0000000000000000000000000ed030828cee7e84ff0031a400541190f813e27e10bd49e61659caec8116d1b022087bda184c51666e470ba12c076d598a0b3d2f77d34c2f83f75a118cfc025392259a2d23d2a4fc31819478afec7be9500d4223f000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000034630b8a0000000000000000000000000000000000000000000000000000000000051e9245200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000003050100000000000000000000000000000000000000000000000000000000b2d05e0000000000000000000000000000000000000000000000000000003691d6afc0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001e00000000000000000000000003189fcb2640bf3c0477b3fbad473af32d8f851e13fccc24b6323e01e8e7a0a504f44f93c434c6b7fd1892f328e1094c5572192086f8abd74e555f875c9a31ccf2855b0a50159a20144955005052f064feac5d5bc000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000034630b8a00000000000000000000000000000000000000000000000000000000000ced5b50200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000003050100000000000000000000000000000000000000000000000000000000b2d05e0000000000000000000000000000000000000000000000000000003691d6afc00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000b2d05e0000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000055d4a80'
+
+w3 = Web3()
+
+with open(ABI_PATH, 'r') as abi_file:
+    abi = json.load(abi_file)
+
+contract = w3.eth.contract(abi=abi)
+
+try:
+    func_obj, func_params = contract.decode_function_input(CALLDATA)
+    decoded_output = {
+        "function_name": func_obj.fn_name,
+        "parameters": func_params
+    }
+    print(decoded_output)
+except Exception as e:
+    print(json.dumps({"error": str(e)}))
