+++ test/engine/fixtures/TestConfigChainDefault.json
@@ -1,0 +1,135 @@
+[
+   {
+      "name": "Expect default config",
+      "steps": [
+         {
+            "time": "0",
+            "tx_id": "1",
+            "tx": {
+               "trace_id": "0",
+               "tx_id": "1",
+               "time": "0",
+               "type": "CREATE_ACCOUNT",
+               "create_account": {
+                  "account": "0x9161ed3bc911f8d3669d1e03da589e4283cb107b",
+                  "signature": {
+                     "signer": "0x9161ed3bc911f8d3669d1e03da589e4283cb107b",
+                     "r": "0x214d080e936476b44894fc9f5ea499963b853e0f093ea78dd9849cf15d46cc12",
+                     "s": "0x3906ccec5342eb6b99868eabbd00fc9648d479eb8ddfc0308a9f35ba28639c98",
+                     "v": 27,
+                     "expiration": "1234",
+                     "nonce": 1234
+                  }
+               }
+            },
+            "tx_abi": "htsA4QAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAACRYe07yRH402adHgPaWJ5Cg8sQewAAAAAAAAAAAAAAAJFh7TvJEfjTZp0eA9pYnkKDyxB7rYo+BmlvW2ROLd6tuu6BF/Qbm5hYgXFfU8pwRx6u2CEJKEC68vqpE0ovs2+84xL/2TjQ3tnwoMJxwhvgcI8XWgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAbAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAE0g==",
+            "tx_data": "0x86db00e1000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000009161ed3bc911f8d3669d1e03da589e4283cb107b0000000000000000000000009161ed3bc911f8d3669d1e03da589e4283cb107bad8a3e06696f5b644e2ddeadbaee8117f41b9b985881715f53ca70471eaed821092840baf2faa9134a2fb36fbce312ffd938d0ded9f0a0c271c21be0708f175a000000000000000000000000000000000000000000000000000000000000001b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004d2",
+            "ret": "",
+            "expectations": [
+               {
+                  "expect": {
+                     "key": "SIMPLE_CROSS_FUTURES_INITIAL_MARGIN",
+                     "sub_key": "0x05",
+                     "value": "0x4e20"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "ADMIN_RECOVERY_ADDRESS",
+                     "sub_key": "0x84b3bc75232c9f880c79efcc5d98e8c6e44f95ae",
+                     "value": "0x01"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "ORACLE_ADDRESS",
+                     "sub_key": "0x47ebfbada4d85dac6b9018c0ce75774556a8243f",
+                     "value": "0x01"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "CONFIG_ADDRESS",
+                     "sub_key": "0xa08ee13480c410de20ea3d126ee2a7daa2a30b7d",
+                     "value": "0x01"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "MARKET_DATA_ADDRESS",
+                     "sub_key": "0x215ec976846b3c68daedf93ba35d725a0e2c98e3",
+                     "value": "0x01"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "ADMIN_FEE_SUB_ACCOUNT_ID",
+                     "value": "0x00"
+                  },
+                  "name": "ExConfig"
+               },
+               {
+                  "expect": {
+                     "key": "INSURANCE_FUND_SUB_ACCOUNT_ID",
+                     "value": "0x00"
+                  },
+                  "name": "ExConfig"
+               },
+               {
+                  "expect": {
+                     "key": "FUNDING_RATE_HIGH",
+                     "sub_key": "0x05",
+                     "value": "0xc350"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "FUNDING_RATE_LOW",
+                     "sub_key": "0x04",
+                     "value": "0xffff3cb0"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "FUTURES_MAKER_FEE_MINIMUM",
+                     "sub_key": "0x04",
+                     "value": "0xffffffe2"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "FUTURES_TAKER_FEE_MINIMUM",
+                     "sub_key": "0x04",
+                     "value": "0x8c"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "OPTIONS_MAKER_FEE_MINIMUM",
+                     "sub_key": "0x04",
+                     "value": "0xffffffe2"
+                  },
+                  "name": "ExConfig2D"
+               },
+               {
+                  "expect": {
+                     "key": "OPTIONS_TAKER_FEE_MINIMUM",
+                     "sub_key": "0x04",
+                     "value": "0x78"
+                  },
+                  "name": "ExConfig2D"
+               }
+            ]
+         }
+      ]
+   }
+]
