+++ test/engine/util.ts
@@ -1,6 +1,8 @@
 import { utils } from "ethers"
-import { Asset } from "./TestEngineTypes"
+import { Asset } from "./types"
 import { CurrencyToEnum, KindToEnum } from "./enums"
+import { TestCase } from "./types"
+import * as fs from "fs"
 
 function getLSB(val: bigint, shift: number) {
   return Number((val >> BigInt(shift)) & BigInt(0xff))
@@ -69,4 +71,54 @@ export const toAssetID = ({ kind, underlying, quote, expiration, strike_price }:
 
 export function hex32(val: string | Uint8Array) {
   return utils.hexZeroPad(utils.hexValue(val), 32)
+}
+
+/**
+ * Scale a string representing a bigint from one decimal precision to another.
+ * @param {string} valueStr - The original value as a string.
+ * @param {number} currentDecimals - The current number of decimals.
+ * @param {number} newDecimals - The desired number of decimals.
+ * @returns {string} - The scaled value as a string.
+ */
+export function scaleBigInt(valueStr: string, currentDecimals: number, newDecimals: number): string {
+  const value = BigInt(valueStr)
+
+  let scaledValue: bigint
+
+  if (currentDecimals < newDecimals) {
+    // Scale up by multiplying by 10^(newDecimals - currentDecimals)
+    scaledValue = value * BigInt(10 ** (newDecimals - currentDecimals))
+  } else if (currentDecimals > newDecimals) {
+    // Scale down by dividing by 10^(currentDecimals - newDecimals)
+    scaledValue = value / BigInt(10 ** (currentDecimals - newDecimals))
+  } else {
+    // No scaling needed
+    scaledValue = value
+  }
+
+  return scaledValue.toString()
+}
+
+export function parseTestsFromFile(filePath: string): TestCase[] {
+  try {
+    // Read the JSON file
+    const data = fs.readFileSync(filePath, "utf8")
+
+    // Parse the JSON data into an array of Test objects
+    try {
+      const tests = JSON.parse(data) as TestCase[]
+      return tests
+    } catch (error) {
+      console.error("Failed to parse JSON:", error)
+      return []
+    }
+  } catch (err) {
+    console.log(`Error reading file from disk: ${err}`)
+    return []
+  }
+}
+
+// Get all json files in the test fixtures directory
+export function getTestFixtures(dir: string): string[] {
+  return fs.readdirSync(dir).filter((file) => file.endsWith(".json"))
 }
